Address comments and improve code quality
