+++ src/hooks/useBestAPY.ts
@@ -1,15 +1,17 @@
 import { Pair, Percent } from '@swapr/sdk'
 import { useMemo } from 'react'
-import { DateTime, Duration } from 'luxon'
 import { useQuery, gql } from '@apollo/client'
-import { getBestApyPairCampaign } from '../utils/liquidityMining'
+import {
+  getBestApyPairCampaign,
+  getRewardTokenAddressFromPair,
+  getLowerTimeLimit,
+  SubgraphPair,
+} from '../utils/liquidityMining'
 import { useActiveWeb3React } from '.'
-import { SubgraphLiquidityMiningCampaign } from '../apollo'
 import { useAllTokensFromActiveListsOnCurrentChain } from '../state/lists/hooks'
 import { useNativeCurrency } from './useNativeCurrency'
 import { useKpiTokens } from './useKpiTokens'
-import { SubgraphToken } from '../utils'
-import { getTokenAmount, getPairWithLiquidityMiningCampaign } from '../utils/index'
+import { getPairWithLiquidityMiningCampaign } from '../utils/liquidityMining'
 
 const QUERY = gql`
   query($lowerTimeLimit: BigInt!, $pairAddress: ID) {
@@ -54,18 +56,6 @@ const QUERY = gql`
   }
 `
 
-interface SubgraphPair {
-  address: string
-  reserve0: string
-  reserve1: string
-  reserveNativeCurrency: string
-  reserveUSD: string
-  totalSupply: string
-  token0: SubgraphToken
-  token1: SubgraphToken
-  liquidityMiningCampaigns: SubgraphLiquidityMiningCampaign[]
-}
-
 interface QueryResult {
   pair: SubgraphPair
 }
@@ -79,16 +69,7 @@ export function useBestAPY(
   const { chainId } = useActiveWeb3React()
   const tokensInCurrentChain = useAllTokensFromActiveListsOnCurrentChain()
   const nativeCurrency = useNativeCurrency()
-
-  const memoizedLowerTimeLimit = useMemo(
-    () =>
-      Math.floor(
-        DateTime.utc()
-          .minus(Duration.fromObject({ days: 150 }))
-          .toSeconds()
-      ),
-    []
-  )
+  const memoizedLowerTimeLimit = useMemo(() => getLowerTimeLimit(), [])
 
   const { loading, data, error } = useQuery<QueryResult>(QUERY, {
     variables: {
@@ -98,11 +79,7 @@ export function useBestAPY(
   })
 
   const rewardTokenAddresses = useMemo(() => {
-    return !data
-      ? []
-      : data.pair.liquidityMiningCampaigns.flatMap(campaign =>
-          campaign.rewards.map(reward => reward.token.address.toLowerCase())
-        )
+    return !data ? [] : getRewardTokenAddressFromPair(data.pair)
   }, [data])
 
   const { loading: loadingKpiTokens, kpiTokens } = useKpiTokens(rewardTokenAddresses)
@@ -110,19 +87,12 @@ export function useBestAPY(
   if (loadingKpiTokens || loading) return { loading: true }
   if (!chainId || error || !data) return { loading: false }
 
-  const { reserveNativeCurrency, totalSupply, token0, token1, reserve0, reserve1, liquidityMiningCampaigns } = data.pair
-
-  const tokenAmountA = getTokenAmount({ token: token0, tokensInCurrentChain, chainId, reserve: reserve0 })
-  const tokenAmountB = getTokenAmount({ token: token1, tokensInCurrentChain, chainId, reserve: reserve1 })
-
   const newPair = getPairWithLiquidityMiningCampaign({
-    pair: new Pair(tokenAmountA, tokenAmountB),
+    rawPair: data.pair,
     chainId,
-    totalSupply,
-    reserveNativeCurrency,
     kpiTokens,
     nativeCurrency,
-    liquidityMiningCampaigns,
+    tokensInCurrentChain,
   })
 
   return {
